fix: 撤回多余文档, SOCKS 测试并入已有文件
